Add CITATION.cff
The CFF file contents were entered based on the information in the
article (author list) and the data repository (keywords, identifier).

CFF initializer was used to simplify file creation:
https://citation-file-format.github.io/cff-initializer-javascript

The identifier used by the data repository is a hdl.net Handle.
Although it resolves successfuly through doi.org, the CFF validator
thought it was not a correct DOI, so it is listed as other.